Fix numlockx use so the bright blue LED indicates the UNwanted state.
Uses xdotool to force the key and indicator out of sync so it's lit when
NumLock is turned off and unlit when it's turned on.

This works because xdotool sets NumLock on the XTEST virtual keyboard
and the key state change bubbles up to the master keyboard but the LED
state change remains contained on the virtual keyboard, throwing the
two out of sync on the physical keyboard.